Move Digitizers to plugins
