+++ SimCalorimetry/HGCalSimProducers/plugins/HGCHEfrontDigitizer.h
